refactor(slice-45): extract pipeline search and signature helpers
Keep phase/status helpers in orchestrator so mega-suite storage patches
stay valid while splitting search and resume-signature concerns (DECISIONS #149).
